move projectRootHeader to PrimaryPane
