+++ src/components/PrimaryPanes/index.js
@@ -27,6 +27,8 @@ import type { SourcesMapByThread } from "../../reducers/types";
 import type { SelectedPrimaryPaneTabType } from "../../selectors";
 import type { Thread } from "../../types";
 
+import Svg from "../shared/Svg";
+
 import "./Sources.css";
 
 type State = {
@@ -42,6 +44,7 @@ type Props = {
   setPrimaryPaneTab: typeof actions.setPrimaryPaneTab,
   setActiveSearch: typeof actions.setActiveSearch,
   closeActiveSearch: typeof actions.closeActiveSearch,
+  clearProjectDirectoryRoot: typeof actions.clearProjectDirectoryRoot,
   threads: Thread[]
 };
 
@@ -98,6 +101,30 @@ class PrimaryPanes extends Component<Props, State> {
     ];
   }
 
+  renderProjectRootHeader() {
+    const { projectRoot } = this.props;
+
+    if (!projectRoot) {
+      return null;
+    }
+
+    const rootLabel = projectRoot.split("/").pop();
+
+    return (
+      <div key="root" className="sources-clear-root-container">
+        <button
+          className="sources-clear-root"
+          onClick={() => this.props.clearProjectDirectoryRoot()}
+          title={L10N.getStr("removeDirectoryRoot.label")}
+        >
+          <Svg name="home" />
+          <Svg name="breadcrumb" />
+          <span className="sources-clear-root-label">{rootLabel}</span>
+        </button>
+      </div>
+    );
+  }
+
   renderThreadSources() {
     return this.props.threads.map(({ actor }) => (
       <SourcesTree thread={actor} key={actor} />
@@ -123,7 +150,10 @@ class PrimaryPanes extends Component<Props, State> {
           })}
           hasFocusableContent
         >
-          <div>{this.renderThreadSources()}</div>
+          <div>
+            {this.renderProjectRootHeader()}
+            {this.renderThreadSources()}
+          </div>
           <Outline
             alphabetizeOutline={this.state.alphabetizeOutline}
             onAlphabetizeClick={this.onAlphabetizeClick}
@@ -147,7 +177,8 @@ const connector = connect(
   {
     setPrimaryPaneTab: actions.setPrimaryPaneTab,
     setActiveSearch: actions.setActiveSearch,
-    closeActiveSearch: actions.closeActiveSearch
+    closeActiveSearch: actions.closeActiveSearch,
+    clearProjectDirectoryRoot: actions.clearProjectDirectoryRoot
   }
 );
